requirements.txt
openai-whisper
streamlit
sounddevice
scipy
numpy
jiwer
langdetect
 